modified sign-up: no verification applied yet
