+++ src/app/sign-up/otp/[phone]/otp-submission.js
@@ -1,13 +1,12 @@
-"use client";
-import React, { useState, useRef } from "react";
+import React, { useState, useRef, useEffect } from "react";
 
-export default function Otp({ onComplete }) {
+const Otp = ({ onComplete, onResend }) => {
   const [otp, setOtp] = useState(Array(6).fill(""));
   const [resendCountdown, setResendCountdown] = useState(30);
   const buttonDisabled = otp.some((value) => value === "");
   const inputRefs = useRef([]);
 
-  React.useEffect(() => {
+  useEffect(() => {
     const timer = setInterval(() => {
       setResendCountdown((prev) => (prev > 0 ? prev - 1 : 0));
     }, 1000);
@@ -17,7 +16,7 @@ export default function Otp({ onComplete }) {
   const handleChange = (event, index) => {
     const { value } = event.target;
     if (value.length === 1 && index < 5) {
-      inputRefs.current[index + 1].focus();
+      inputRefs.current[index + 1]?.focus();
     }
     setOtp((prev) => {
       const newOtp = [...prev];
@@ -27,11 +26,12 @@ export default function Otp({ onComplete }) {
   };
 
   const handleConfirm = () => {
-    onComplete();
+    onComplete(otp.join(""));
   };
 
   const handleResendCode = () => {
     setResendCountdown(30);
+    onResend();
   };
 
   return (
@@ -45,6 +45,7 @@ export default function Otp({ onComplete }) {
               className="flex-1 text-center rounded border-2 border-accent2 w-12 h-12 text-primary text-2xl otp"
               type="text"
               maxLength={1}
+              value={otp[index]}
               onChange={(event) => handleChange(event, index)}
             />
           ))}
@@ -67,25 +68,22 @@ export default function Otp({ onComplete }) {
           )}
         </div>
         <div style={{ marginTop: "9rem" }}>
-          {buttonDisabled ? (
-            <button
-              className="bg-accent px-6 py-3 w-full rounded-3xl text-sm font-medium text-primary text-opacity-40"
-              type="button"
-              disabled={buttonDisabled}
-            >
-              CONFIRM
-            </button>
-          ) : (
-            <button
-              className="bg-primary px-6 py-3 w-full rounded-3xl text-sm font-medium text-white"
-              type="button"
-              onClick={handleConfirm}
-            >
-              CONFIRM
-            </button>
-          )}
+          <button
+            className={`px-6 py-3 w-full rounded-3xl text-sm font-medium ${
+              buttonDisabled
+                ? "bg-accent text-primary text-opacity-40"
+                : "bg-primary text-white"
+            }`}
+            type="button"
+            onClick={handleConfirm}
+            disabled={buttonDisabled}
+          >
+            CONFIRM
+          </button>
         </div>
       </div>
     </div>
   );
-}
+};
+
+export default Otp;
